+++ frontend/src/app/(routes)/(dashboard)/dashboard/courses/[id]/page.tsx
@@ -1,4 +1,4 @@
-import {TabsContent} from '@radix-ui/react-tabs'
+import { TabsContent } from '@/components/ui/tabs'
 import ChatComponent from '@/components/chat'
 import { getCourse } from '@/lib/courses'
